# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: gin-template API
  version: 0.1.0
servers:
  - url: /api/v1
components:
  responses:
    ProblemResponse:
      description: Problem
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Problem' }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponseBase:
      type: object
      required: [status, msg]
      properties:
        status: { type: integer, description: 0 表示成功；100000-199999 为系统级状态码；200000-299999 为业务级状态码，按 biz 模块分段 }
        msg: { type: string }
    Problem:
      type: object
      required: [status, msg, details, data]
      properties:
        status: { type: integer }
        msg: { type: string }
        details: { type: string }
        data:
          nullable: true
    User:
      type: object
      required: [uid, username, email, display_name, role, status, email_verified, ctime, utime]
      properties:
        uid: { type: integer, format: int64 }
        username: { type: string }
        email: { type: string }
        display_name: { type: string }
        role: { type: string }
        status: { type: string }
        email_verified: { type: boolean }
        ctime: { type: integer, format: int64, description: Unix 秒级时间戳 }
        utime: { type: integer, format: int64, description: Unix 秒级时间戳 }
    AuthResponse:
      type: object
      required: [access_token, expires_at, user]
      properties:
        access_token: { type: string }
        expires_at: { type: integer, format: int64, description: Unix 秒级时间戳 }
        user: { $ref: '#/components/schemas/User' }
    AuthEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/AuthResponse' }
    LoginRequest:
      type: object
      required: [identifier, password]
      properties:
        identifier: { type: string }
        password: { type: string }
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username: { type: string }
        email: { type: string, format: email }
        password: { type: string }
        display_name: { type: string }
        turnstile_token: { type: string }
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
        debug_token: { type: string }
    MessageEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/MessageResponse' }
    TokenRequest:
      type: object
      required: [token]
      properties:
        token: { type: string }
        new_password: { type: string }
        email: { type: string, format: email }
        turnstile_token: { type: string }
...
    File:
      type: object
      required: [id, uid, original_name, storage_name, content_type, size, path, ctime, utime]
      properties:
        id: { type: string }
        uid: { type: integer, format: int64 }
        original_name: { type: string }
        storage_name: { type: string }
        content_type: { type: string }
        size: { type: integer }
        path: { type: string }
        ctime: { type: integer, format: int64, description: Unix 秒级时间戳 }
        utime: { type: integer, format: int64, description: Unix 秒级时间戳 }
    FileListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/File' }
        total: { type: integer }
    FileEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/File' }
    FileListEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/FileListResponse' }
...
paths:
  /auth/register:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthEnvelope' }
        default:
          $ref: '#/components/responses/ProblemResponse'
  /auth/login:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthEnvelope' }
        default:
          $ref: '#/components/responses/ProblemResponse'
  /auth/logout:
    post:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageEnvelope' }
        default:
          $ref: '#/components/responses/ProblemResponse'
  /auth/refresh:
    post:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthEnvelope' }
        default:
          $ref: '#/components/responses/ProblemResponse'
  /auth/password/request-reset:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TokenRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageEnvelope' }
        default:
          $ref: '#/components/responses/ProblemResponse'
  /auth/password/reset:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TokenRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageEnvelope' }
        default:
          $ref: '#/components/responses/ProblemResponse'
  /auth/email/send-verification:
    post:
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageEnvelope' }
        default:
          $ref: '#/components/responses/ProblemResponse'
  /auth/email/verify:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TokenRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageEnvelope' }
        default:
          $ref: '#/components/responses/ProblemResponse'
...
  /files:
    get:
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileListEnvelope' }
        default:
          $ref: '#/components/responses/ProblemResponse'
  /files/search:
    get:
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileListEnvelope' }
        default:
          $ref: '#/components/responses/ProblemResponse'
  /files/{id}:
    get:
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileEnvelope' }
        default:
          $ref: '#/components/responses/ProblemResponse'
    delete:
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageEnvelope' }
        default:
          $ref: '#/components/responses/ProblemResponse'
  /files/upload:
    post:
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileEnvelope' }
        default:
          $ref: '#/components/responses/ProblemResponse'
...